requirements.txt
google-genai
python-dotenv
RealtimeSTT
elevenlabs
feedparser
requests
beautifulsoup4
wikipedia
